+++ src/DataFilters/Grammar/Syntax/FilterExpression.cs
@@ -19,7 +19,6 @@
         /// <returns>
         /// <c>true</c> if <paramref name="other"/> is "equivalent" to the current instance.
         /// </returns>
-        /// <inheritdoc/>
         public virtual bool IsEquivalentTo(FilterExpression other)
             => Equals(other) || Equals((other as ISimplifiable)?.Simplify());